+++ src/Bedrock.Framework.Experimental/Client/ClientExtensions.cs
@@ -6,7 +6,7 @@ namespace Bedrock.Framework
         public static ClientBuilder UseConnectionPooling(this ClientBuilder builder)
         {
             // TODO right now, this must be called after UseSockets
-            builder.ConnectionFactory = new ConnectionPoolingFactory(new HttpConnectionPool(builder.ConnectionFactory));
+            builder.ConnectionFactory = new ConnectionPoolingFactory(new ConnectionPool(builder.ConnectionFactory));
             return builder;
         }
     }
