+++ src/Bedrock.Framework.Experimental/Transports/Pooling/IEndPointPool.cs
@@ -1,0 +1,12 @@
+using System.Threading;
+using System.Threading.Tasks;
+using Microsoft.AspNetCore.Connections;
+
+namespace Bedrock.Framework
+{
+    internal interface IEndPointPool
+    {
+        ValueTask ReturnAsync(ConnectionContext context);
+        ValueTask<ConnectionContext> GetConnectionAsync(CancellationToken cancellationToken = default);
+    }
+}
